LayoutTests/fast/repaint/outline-auto-with-default-width-and-offset-repaint-expected.txt

PASS window.internals.repaintRectsAsText().indexOf('4 6') is not -1
PASS window.internals.repaintRectsAsText().indexOf('4 46') is not -1
PASS successfullyParsed is true

TEST COMPLETE
